// src/app/modules/OfferedCourse/offeredCourse.interface.ts
import { Types } from 'mongoose'

export type TDays = 'Sat' | 'Sun' | 'Mon' | 'Tue' | 'Wed' | 'Thu' | 'Fri'

export type TOfferedCourse = {
    semesterRegistration: Types.ObjectId
    academicSemester?: Types.ObjectId
    academicFaculty: Types.ObjectId
    academicDepartment: Types.ObjectId
    course: Types.ObjectId
    faculty: Types.ObjectId
    maxCapacity: number
    section: string
    days: TDays[]
    startTime: string
    endTime: string
}